+++ libre_pvz/src/core.rs
@@ -16,14 +16,9 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-//! librePvZ: game logic implementation.
+//! Core subsystems: contributes core functionalities to the game implementation.
 
-#![warn(missing_docs)]
-#![warn(missing_debug_implementations)]
-
-pub use libre_pvz_resources as resources;
-pub use libre_pvz_animation as animation;
-
-pub mod core;
-pub mod scene;
-pub mod diagnostics;
+pub mod behaviour;
+pub mod geometry;
+pub mod kinematics;
+pub mod projectile;