+++ libre_pvz/src/main.rs
@@ -17,22 +17,29 @@
  */
 
 use bevy::prelude::*;
+use bevy_egui::EguiPlugin;
 use libre_pvz::resources::bevy::{Animation, AnimationLoader};
 use libre_pvz::animation::AnimationPlugin;
-use libre_pvz::almanac::AlmanacPlugin;
+use libre_pvz::diagnostics::BoundingBoxPlugin;
+// use libre_pvz::scene::almanac::AlmanacPlugin;
+use libre_pvz::scene::lawn::LawnPlugin;
 
 fn main() {
-    let anim_name = std::env::args().into_iter().nth(1)
-        .unwrap_or_else(|| "Peashooter.anim".to_string());
+    // let anim_name = std::env::args().into_iter().nth(1)
+    //     .unwrap_or_else(|| "Peashooter.anim".to_string());
 
     App::new()
-        .insert_resource(AlmanacPlugin::window_descriptor())
+        .insert_resource(LawnPlugin::window_descriptor())
+        // .insert_resource(AlmanacPlugin::window_descriptor())
         .add_plugins(DefaultPlugins)
+        .add_plugin(EguiPlugin)
+        .add_plugin(BoundingBoxPlugin)
         .add_plugin(AnimationPlugin)
         .add_asset::<Animation>()
         .init_asset_loader::<AnimationLoader>()
         .add_startup_system(setup_camera)
-        .add_plugin(AlmanacPlugin::new(anim_name))
+        .add_plugin(LawnPlugin)
+        // .add_plugin(AlmanacPlugin::new(anim_name))
         .run();
 }