+++ libre_pvz/src/scene.rs
@@ -16,14 +16,8 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-//! librePvZ: game logic implementation.
+//! Different scenes of the game.
 
-#![warn(missing_docs)]
-#![warn(missing_debug_implementations)]
-
-pub use libre_pvz_resources as resources;
-pub use libre_pvz_animation as animation;
-
-pub mod core;
-pub mod scene;
-pub mod diagnostics;
+pub mod loading;
+pub mod almanac;
+pub mod lawn;